+++ src/generatorV3.ts
@@ -199,6 +199,7 @@ export class ApiGenerator {
       naming,
       httpClientConfig: this.httpClientConfig,
       controllers,
+      tags: apiDocs.tags || [],
       describe,
       typeImportCandidates,
       typeMode,